* commit 'c2bf2f0f4dbf9ae01ab0eb944e0185c733b5e225': (21 commits)
  ZP-1413 Turn php dependencies around. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1412 Add ignore_missing_attachments option to mapi_inetmapi_imtoinet.
  ZP-1411 Unset undefined properties for tasks.
  ZP-1410 Dot at the end of sentences. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1410 IMAP backend to provide user details for CalDAV backend. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1408 Don't log a warning when a user does not enter a subject. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1372 PHP doc fix - added a dot at the end of the sentence.
  ZP-1372 Created getter ZPush::GetDeviceManager()->GetDevid() , resumed SYNC_STATUS_INVALIDSYNCKEY for triggering a folder resync. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-250 Improve log messages in readPropStream.
  ZP-1401 Sync Kopano contacts with gab2contacts.
  ZP-1400 Picture not saved for a contact created on mobile.
  ZP-1399 Wrong state and log folder permissions for RHEL based systems.
  ZP-1398 Do not call $device->GetHierarchyCache() in loop.
  ZP-1372 Removed logging deviceid. Code style changes.
  ZP-1396 Only a meeting if meetingstatus says so. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1369 Get foldername from MAPI when impersonating.
  ZP-1372 Moved 'Not Triggering folder re-sync' message to INFO level. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1372 Folder re-sync is triggered on deletions ratio threshold. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1386 Update mimedecode include. Released under the Affero GNU General Public License (AGPL) version 3.
  ZP-1391 Support disabling multiple IMAP auth methods, log the disabled methods. Released under the Affero GNU General Public License (AGPL) version 3.
  ...

+++ src/lib/utils/utils.php
@@ -894,14 +894,15 @@ class Utils {
             $perm_dir = stat($dir);
             $perm_file = stat($file);
 
-            if ($perm_file['uid'] == 0 && $perm_dir['uid'] == 0) {
+            if ($perm_file['uid'] == 0 && $perm_dir['uid'] == 0 && $perm_dir['gid'] == 0) {
                 unlink($file);
-                throw new FatalException("FixFileOwner: $dir must be owned by the nginx/apache/php user instead of root");
+                throw new FatalException("FixFileOwner: $dir must be owned by the nginx/apache/php user instead of root for debian based systems and by root:z-push for RHEL-based systems");
             }
 
             if($perm_dir['uid'] !== $perm_file['uid'] || $perm_dir['gid'] !== $perm_file['gid']) {
                 chown($file, $perm_dir['uid']);
                 chgrp($file, $perm_dir['gid']);
+                chmod($file, 0664);
             }
         }
         return true;
@@ -1361,7 +1362,7 @@ class Utils {
         if (!$rawheaders) {
             return;
         }
-        $message->headers["subject"] = Utils::convertRawHeader2Utf8($rawheaders["subject"], $message->headers["subject"]);
+        $message->headers["subject"] = isset($message->headers["subject"]) ? Utils::convertRawHeader2Utf8($rawheaders["subject"], $message->headers["subject"]) : "";
         $message->headers["from"] = Utils::convertRawHeader2Utf8($rawheaders["from"], $message->headers["from"]);
     }
 }
